shared/idmap: Move ByHostID back into set
